+++ flexapp-vuln-scanner-go/cmd/tray/main_windows.go
@@ -14,6 +14,7 @@ import (
 	_ "embed"
 	"fmt"
 	"image/png"
+	"net/http"
 	"os"
 	"os/exec"
 	"path/filepath"
@@ -63,6 +64,8 @@ type app struct {
 	logOffset    int64
 	logRuneCount int
 
+	pickerServer *http.Server
+
 	mu  sync.Mutex
 	cmd *exec.Cmd
 }
@@ -110,6 +113,8 @@ func main() {
 	if err := a.buildNotifyIcon(); err != nil {
 		warnDialog("build tray icon", err)
 	}
+
+	a.pickerServer = a.startPickerServer(currentPickerAddr(installDir))
 
 	a.start()
 	a.mw.Run()
@@ -419,6 +424,9 @@ func (a *app) restart() {
 
 func (a *app) quit() {
 	a.stop()
+	if a.pickerServer != nil {
+		a.pickerServer.Close()
+	}
 	if a.notifyIcon != nil {
 		a.notifyIcon.Dispose()
 	}